FST 83 - refactor
